Desktop infoSlider implementation
